build emacs

+++ bin/build-emacs.sh
@@ -1,11 +1,17 @@
 #!/bin/sh
-rm -rf ~/Library/Caches/Homebrew/emacs-plus@30--git
-ln -sf ~/repos/emacs ~/Library/Caches/Homebrew/emacs-plus@30--git
+if [ `uname -s` = "Linux" ] ; then
+    cd /home/jixiuf/emacs; git fetch --all;git pull
+    sudo emerge -av emacs
+else
+    rm -rf ~/Library/Caches/Homebrew/emacs-plus@30--git
+    ln -sf ~/repos/emacs ~/Library/Caches/Homebrew/emacs-plus@30--git
 
-# cd /usr/local/Homebrew/Library/Taps/d12frosted/homebrew-emacs-plus/ ;git diff >~/.emacs.d/bin/emacs.patch
-# cd /usr/local/Homebrew/Library/Taps/d12frosted/homebrew-emacs-plus/
-# patch -t -N -p1 <~/.emacs.d/bin/emacs.patch
-# cd -
+    # cd /usr/local/Homebrew/Library/Taps/d12frosted/homebrew-emacs-plus/ ;git diff >~/.emacs.d/bin/emacs.patch
+    # cd /usr/local/Homebrew/Library/Taps/d12frosted/homebrew-emacs-plus/
+    # patch -t -N -p1 <~/.emacs.d/bin/emacs.patch
+    # cd -
 
-brew reinstall --with-native-comp --with-xwidgets --with-poll emacs-plus@30
-# cd /usr/local/Homebrew/Library/Taps/d12frosted/homebrew-emacs-plus/;git stash ;git stash drop
+    brew reinstall --with-native-comp --with-xwidgets --with-poll emacs-plus@30
+    # cd /usr/local/Homebrew/Library/Taps/d12frosted/homebrew-emacs-plus/;git stash ;git stash drop
+
+fi
